requirements.txt
amqp==5.1.1
anyio==3.6.1
apt-clone==0.2.1
apturl==0.5.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.2
asgiref==3.5.0
asttokens==2.0.5
attrs==21.4.0
autopep8==1.6.0
Babel==2.10.3
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.8.2
billiard==3.6.4.0
bleach==5.0.1
blessed==1.19.1
blinker==1.4
Brlapi==0.7.0
celery==5.2.7
certifi==2022.6.15
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.0.12
chrome-gnome-shell==0.0.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
cryptography==2.8
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.2.16
debugpy==1.6.2
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
Deprecated==1.2.13
distlib==0.3.4
distro==1.4.0
Django==4.0.3
django-aloha-edit==0.3.3
django-compat==1.0.15
django-cors-headers==3.11.0
django-filter==21.1
django-picklefield==3.0.1
django-q==1.3.9
django-quill-editor==0.1.40
django-resized==1.0.0
django4-background-tasks==1.2.7
djangorestframework==3.13.1
djangorestframework-recursive==0.1.2
djangorestframework-timed-auth-token==1.3.0
drf-nested-routers==0.93.4
entrypoints==0.3
executing==0.8.3
fastjsonschema==2.16.1
filelock==3.7.1
fonttools==4.36.0
grpcio==1.16.1
gunicorn==20.1.0
httplib2==0.14.0
idna==3.3
ifaddr==0.1.6
IMDbPY==6.8
importlib-metadata==4.12.0
importlib-resources==5.9.0
install==1.3.5
ipykernel==6.15.1
ipython==8.4.0
ipython-genutils==0.2.0
jedi==0.18.1
Jinja2==3.1.2
json5==0.9.8
jsonschema==4.7.2
jupyter-client==7.3.4
jupyter-core==4.11.1
jupyter-server==1.18.1
jupyterlab==3.4.4
jupyterlab-pygments==0.2.2
jupyterlab-server==2.15.0
kazam==1.4.5
keyring==18.0.1
kiwisolver==1.4.4
kombu==5.2.4
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
louis==3.12.0
lxml==4.9.1
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline==0.1.3
mistune==0.8.4
mysqlclient==2.1.0
nbclassic==0.4.3
nbclient==0.6.6
nbconvert==6.5.0
nbformat==5.4.0
nemo-emblems==5.0.0
nest-asyncio==1.5.5
netaddr==0.7.19
netifaces==0.10.4
notebook==6.4.12
notebook-shim==0.1.0
numpy==1.23.2
oauthlib==3.1.0
onboard==1.4.1
packaging==20.3
PAM==0.4.2
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.9.0
pexpect==4.6.0
pickleshare==0.7.5
Pillow==9.1.0
platformdirs==2.5.2
prometheus-client==0.14.1
prompt-toolkit==3.0.29
protobuf==3.6.1
psutil==5.5.1
ptyprocess==0.7.0
pure-eval==0.2.2
pycairo==1.16.2
pycodestyle==2.8.0
pycparser==2.21
pycrypto==2.6.1
pycups==1.9.73
pycurl==7.43.0.2
PyGithub==1.55
Pygments==2.12.0
PyGObject==3.36.0
PyICU==2.4.2
pyinotify==0.9.6
PyJWT==2.4.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.6
pyparted==3.11.2
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
pyRFC3339==1.1
pyrsistent==0.18.1
python-apt==2.0.0+ubuntu0.20.4.5
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python-dotenv==0.20.0
python-magic==0.4.16
python-xapp==2.2.1
python-xlib==0.23
pytz==2022.1
pyxdg==0.26
PyYAML==5.3.1
pyzmq==23.2.0
redis==3.5.3
reportlab==3.6.10
requests==2.28.0
requests-file==1.4.3
requests-unixsocket==0.2.0
screenpen==0.1.14
SecretStorage==2.3.1
Send2Trash==1.8.0
setproctitle==1.1.10
simplejson==3.16.0
six==1.16.0
sniffio==1.2.0
soupsieve==1.9.5
sqlparse==0.4.2
stack-data==0.3.0
systemd-python==234
terminado==0.15.0
tinycss2==1.0.2
tldextract==2.2.1
toml==0.10.2
tornado==6.2
traitlets==5.3.0
ubuntu-drivers-common==0.0.0
ufw==0.36
Unidecode==1.1.1
urllib3==1.26.9
vine==5.0.0
wadllib==1.3.3
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.3
whitenoise==6.0.0
wrapt==1.14.1
xkit==0.0.0
youtube-dl==2021.4.26
zipp==3.8.1
